fix Visual Studiuo x64 warnings
Author: Chiayang Tsai <chiayang.tsai@mediatek.com>


git-svn-id: svn://hevc.kw.bbc.co.uk/svn/jctvc-tmuc/branches/HM-4.1-dev@1466 ce97f6c2-8f86-0410-b44d-af84b13ec34f

+++ source/Lib/TLibEncoder/TEncAdaptiveLoopFilter.cpp
@@ -3359,7 +3359,7 @@ Void TEncAdaptiveLoopFilter::xSetCUAlfCtrlFlags_qc(UInt uiAlfCtrlDepth, TComPicY
       xSetCUAlfCtrlFlag_qc(pcCU, 0, 0, uiAlfCtrlDepth, pcPicOrg, pcPicDec, pcPicRest, ruiDist, cAlfCUCtrlParam.alf_cu_flag);
     }
 
-    cAlfCUCtrlParam.num_alf_cu_flag = cAlfCUCtrlParam.alf_cu_flag.size();
+    cAlfCUCtrlParam.num_alf_cu_flag = (UInt)cAlfCUCtrlParam.alf_cu_flag.size();
   }
   else
   {
@@ -5466,7 +5466,7 @@ Void TEncAdaptiveLoopFilter::transferCtrlFlagsToAlfParam(std::vector<AlfCUCtrlIn
         }
       }
     }
-    cSliceCUCtrlParam.num_alf_cu_flag = cSliceCUCtrlParam.alf_cu_flag.size();
+    cSliceCUCtrlParam.num_alf_cu_flag = (UInt)cSliceCUCtrlParam.alf_cu_flag.size();
   }
 }
